+++ src/loggers/utils.js
@@ -1,0 +1,9 @@
+import _ from 'lodash';
+
+export function repeat(char, count) {
+  return _.chain()
+    .range(count)
+    .map(() => char)
+    .value()
+    .join('');
+}
